riffing on the nav again
