+++ src/SideNav/SideNav.tsx
@@ -1,19 +1,11 @@
 import React from 'react'
-import { Box, Link } from 'rebass'
-import { useHistory } from 'react-router-dom'
-import { Inbox } from 'react-feather'
+import { Box } from 'rebass'
 
 import Teams from './Teams'
 
 import { Logo } from 'components'
 
 export const SideNav: React.FC = ({ children }) => {
-  const history = useHistory()
-  const navigate = (to: string) => (ev: React.MouseEvent) => {
-    ev.preventDefault()
-    history.push(to)
-  }
-
   return (
     <Box
       as="aside"
@@ -40,21 +32,6 @@ export const SideNav: React.FC = ({ children }) => {
       </Box>
 
       <Box px={3} mb={3}>
-        <Link
-          sx={{
-            alignItems: 'center',
-            display: 'inline-flex',
-            fontSize: 2,
-            mb: 3,
-            textDecoration: 'none',
-          }}
-          color="text"
-          onClick={navigate('/library')}
-          href="#"
-        >
-          <Box as={Inbox} size={[16, 20]} color="muted" mr={2} />
-          Library
-        </Link>
         {children}
       </Box>
     </Box>
